split up angular ket
