+++ src/rydstate/angular/__init__.py
@@ -1,5 +1,6 @@
 from rydstate.angular import utils
-from rydstate.angular.angular_ket import AngularKetDummy, AngularKetFJ, AngularKetJJ, AngularKetLS
+from rydstate.angular.angular_ket import AngularKetFJ, AngularKetJJ, AngularKetLS
+from rydstate.angular.angular_ket_dummy import AngularKetDummy
 from rydstate.angular.angular_state import AngularState
 
 __all__ = [
